\name{get_trT}
\alias{get_trT}
\alias{lndetcheb_in}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Chebyshev Jacobian approximation for spatial regression}
\description{
  
}
\usage{
get_trT(W, q = 5)
lndetcheb_in(trT, alpha)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{W}{}
  \item{q}{}
  \item{trT}{}
  \item{alpha}{}
}
\details{
  
}
\value{
  ~Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
  ...
}
\references{}
\author{}

\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{spatial}

