\name{mcdet_prep}
\alias{mcdet_prep}
\alias{mcdet_in}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Monte Carlo Jacobian approximation for spatial regression}
\description{
  
}
\usage{
mcdet_prep(W, p = 16, m = 30)
mcdet_in(alpha, clx)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{W}{}
  \item{p}{}
  \item{m}{}
  \item{alpha}{}
  \item{clx}{}
}
\details{
  
}
\value{
  ~Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
  ...
}
\references{}
\author{}

\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{spatial}

